Updates for VS projects.

+++ src/util.hpp
@@ -2,6 +2,7 @@
 #define SX_UTIL_HPP
 
 #include <iostream>
+#include <random>
 #include <boost/lexical_cast.hpp>
 #include <boost/algorithm/string.hpp>
 #include <bitcoin/bitcoin.hpp>
@@ -100,7 +101,7 @@ bool read_hd_command_args(int argc, char** argv,
         std::cerr << "Usage: sx hd-priv [--hard] INDEX" << std::endl;
         return false;
     }
-    for (size_t i = 1; i < argc; ++i)
+    for (int i = 1; i < argc; ++i)
     {
         std::string arg = argv[i];
         if (arg == "-h" || arg == "--hard")
